refactor: 最新月予算コンポーネントの公開名を整理する
最新月予算コンポーネントのディレクトリ名をUI向けに揃え、設定画面テストの型名衝突を避けます。
