--- apps/web/src/features/budgets/getLatestMonthlyBudget/index.ts
@@ -1,1 +1,0 @@
-export * from "./LatestMonthlyBudget/LatestMonthlyBudget"
