+++ apps/web/src/features/budgets/latestMonthlyBudget/index.ts
@@ -1,0 +1,1 @@
+export { LatestMonthlyBudget } from "./LatestMonthlyBudget/LatestMonthlyBudget"
